ml_system/client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
